// src/api/generated/schemas/branchRequest.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */
import type { AddressRequest } from './addressRequest';
import type { ContactRequest } from './contactRequest';

export interface BranchRequest {
  documento?: string;
  isMatriz?: boolean;
  /**
     * @minLength 0
     * @maxLength 100
     */
  apelido?: string;
  /**
     * @minLength 0
     * @maxLength 20
     */
  inscricaoEstadual?: string;
  /**
     * @minLength 0
     * @maxLength 20
     */
  inscricaoMunicipal?: string;
  address?: AddressRequest;
  contatos?: ContactRequest[];
}